wip: restructuring files to support multiple targets
