+++ .bazelrc
@@ -3,7 +3,7 @@
 
 ##### WASM #####
 # Use our custom-configured c++ toolchain.
-build:wasm --crosstool_top=//toolchain:emscripten
+build:wasm --crosstool_top=//src/javascript/toolchain:emscripten
 
 # Use --cpu as a differentiator.
 build:wasm --cpu=wasm
@@ -24,7 +24,7 @@ build:wasm --spawn_strategy=local
 
 ##### Pure JS #####
 # Use our custom-configured c++ toolchain.
-build:js --crosstool_top=//toolchain:emscripten
+build:js --crosstool_top=//src/javascript/toolchain:emscripten
 
 # Use --cpu as a differentiator.
 build:js --cpu=wasm
@@ -34,10 +34,10 @@ build:js --host_crosstool_top=@bazel_tools//tools/cpp:toolchain
 
 # These compile flags are active no matter which build mode we are in
 # (dbg vs opt). For flags specific to build mode, see cc_toolchain_config.bzl.
-build:js --cxxopt="-std=c++17"
-build:js --cxxopt="-fno-rtti"
-build:js --cxxopt="-fno-exceptions"
-build:js --cxxopt="-fomit-frame-pointer"
+build:js --host_cxxopt="-std=c++17"
+build:js --host_cxxopt="-fno-rtti"
+build:js --host_cxxopt="-fno-exceptions"
+build:js --host_cxxopt="-fomit-frame-pointer"
 
 # Disable sandbox environment because emsdk caches files by writing to
 # home directory.
