+++ CONTRIBUTORS.md
@@ -1,1 +1,2 @@
 [Phillipp Schoppmann](https://github.com/schoppmp)
+[Nick Angelou](https://github.com/s0l0ist)
