+++ README.md
@@ -4,15 +4,15 @@ Private Set Intersection Cardinality protocol based on ECDH and Bloom Filters.
 ## Compiling and Running
 The only requirement is [Bazel](https://bazel.build). To compile, run
 ```
-bazel build //...
+bazel build //src/cpp/...
 ```
 To run tests, use
 ```
-bazel test //...
+bazel test //src/cpp/...
 ```
 Benchmarks can be run using
 ```
-bazel run -c opt //psi_cardinality:psi_cardinality_benchmark
+bazel run -c opt //src/cpp:psi_cardinality_benchmark
 ```
 
 ## Using the Library
@@ -21,16 +21,16 @@ To use this library in a Bazel project, put the following in your WORKSPACE file
 load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
 
 http_archive(
-   name = "com_github_schoppmp_psi_cardinality",
+   name = "com_github_openmined_psi_cardinality",
    strip_prefix = "psi-cardinality-master",
-   url = "https://github.com/schoppmp/psi-cardinality/archive/master.zip",
+   url = "https://github.com/OpenMined/psi-cardinality/archive/master.zip",
 )
 
-load("@com_github_schoppmp_psi_cardinality//psi_cardinality:preload.bzl", "psi_cardinality_preload")
+load("@com_github_openmined_psi_cardinality//psi_cardinality:preload.bzl", "psi_cardinality_preload")
 
 psi_cardinality_preload()
 
-load("@com_github_schoppmp_psi_cardinality//psi_cardinality:deps.bzl", "psi_cardinality_deps")
+load("@com_github_openmined_psi_cardinality//psi_cardinality:deps.bzl", "psi_cardinality_deps")
 
 psi_cardinality_deps()
 ```
