+++ scripts/em-init.sh
@@ -2,7 +2,7 @@
 set -euo pipefail
 
 # Sets up llvm
-cd submodules/emsdk/ \
-&& ./emsdk install latest \
-&& ./emsdk activate latest \
-&& cd ../../
+cd submodules/emsdk/
+./emsdk install latest
+./emsdk activate latest
+cd ../../
