+++ scripts/em-update.sh
@@ -2,6 +2,6 @@
 set -euo pipefail
 
 # Updates emsdk to the latest version
-cd ./submodules/emsdk/ \
-&& ./emsdk update-tags \
-&& cd ../../
+cd ./submodules/emsdk/
+./emsdk update-tags
+cd ../../
