+++ src/cpp/psi_cardinality_client_test.cpp
@@ -19,7 +19,7 @@
 #include "absl/strings/str_cat.h"
 #include "crypto/ec_commutative_cipher.h"
 #include "gtest/gtest.h"
-#include "psi_cardinality/bloom_filter.h"
+#include "bloom_filter.h"
 #include "rapidjson/document.h"
 #include "rapidjson/stringbuffer.h"
 #include "rapidjson/writer.h"
