+++ src/javascript/toolchain/cc_toolchain_config.bzl
@@ -82,22 +82,18 @@ def _impl(ctx):
                         flags = [
                             "-isystem",
                             "external/emsdk/emsdk/upstream/emscripten/system/include/libcxx",
-                            # "-isystem",
-                            # "external/emsdk/emsdk/upstream/emscripten/system/lib/libcxxabi/include",
+                            "-isystem",
+                            "external/emsdk/emsdk/upstream/emscripten/system/lib/libcxxabi/include",
                             "-isystem",
                             "external/emsdk/emsdk/upstream/emscripten/system/include/compat",
                             "-isystem",
                             "external/emsdk/emsdk/upstream/emscripten/system/include",
                             "-isystem",
                             "external/emsdk/emsdk/upstream/emscripten/system/include/libc",
-                            # "-isystem",
-                            # "external/emsdk/emsdk/upstream/emscripten/system/include/libc/sys",
                             "-isystem",
                             "external/emsdk/emsdk/upstream/emscripten/system/lib/libc/musl/arch/emscripten",
-                            # "-isystem",
-                            # "external/emsdk/emsdk/upstream/emscripten/system/lib/libc/musl/src/internal",
-                            # "-isystem",
-                            # "external/emsdk/emsdk/upstream/emscripten/system/local/include",
+                            "-isystem",
+                            "external/emsdk/emsdk/upstream/emscripten/system/local/include",
                         ],
                     ),
                 ],
@@ -112,9 +108,10 @@ def _impl(ctx):
             flag_groups = [flag_group(flags = ["-DNDEBUG"])],
             with_features = [with_feature_set(features = ["opt"])],
         ),
+        # Overriding to use -O2 instead of -O3 becuase asmjs breaks.
         flag_set(
             actions = all_compile_actions + all_link_actions,
-            flag_groups = [flag_group(flags = ["-g0", "-O3"])],
+            flag_groups = [flag_group(flags = ["-g0", "-O2"])], 
             with_features = [with_feature_set(features = ["opt"])],
         ),
         # Fastbuild (fastbuild)
